[Add]eMenuPosition列挙型を追加
メニューボタンのアニメーションを追加しました。

+++ Sports/Sports/Struct/sXPosition.swift
@@ -6,7 +6,7 @@ import SpriteKit
 struct SXPosition {
     
     var gameWidth : Int = 540
-    var center : Int = 270
+    var center : CGFloat = 270
     
     var parts : [[CGFloat]] = []
